+++ workspace/packages/muse-core/types/am/deleteApp.d.ts
@@ -1,0 +1,7 @@
+declare function _exports(params?: {
+    appName: string;
+    author?: string;
+}): Promise<{
+    error: any;
+}>;
+export = _exports;