+++ workspace/packages/muse-core/types/am/export.d.ts
@@ -1,0 +1,2 @@
+declare function _exports(params?: ExportArgument): Promise<void>;
+export = _exports;