+++ workspace/packages/muse-core/types/data/builders/muse.plugin.d.ts
@@ -1,0 +1,3 @@
+export function match(key: any): boolean;
+export function get(key: any): Promise<Buffer>;
+export function getMuseDataKeysByRawKeys(rawDataType: any, keys: any): any;