+++ workspace/packages/muse-core/types/data/builders/muse.plugins.d.ts
@@ -1,0 +1,3 @@
+export const key: string;
+export function get(): Promise<Buffer[]>;
+export function getMuseDataKeysByRawKeys(rawDataType: any, keys: any): string;