+++ workspace/packages/muse-core/types/pm/deleteRelease.d.ts
@@ -1,0 +1,7 @@
+declare function _exports(params: {
+    pluginName: string;
+    version: string;
+    author?: string;
+    msg?: string;
+}): Promise<any>;
+export = _exports;