+++ workspace/packages/muse-core/types/req/completeRequest.d.ts
@@ -1,0 +1,6 @@
+declare function _exports({ requestId, msg, author }: {
+    requestId: string;
+    author?: string;
+    msg?: string;
+}): request;
+export = _exports;