+++ workspace/packages/muse-core/types/req/getRequest.d.ts
@@ -1,0 +1,2 @@
+declare function _exports(requestId: string): Buffer;
+export = _exports;