+++ workspace/packages/muse-core/types/req/updateStatus.d.ts
@@ -1,0 +1,7 @@
+declare function _exports(params: {
+    requestId: string;
+    status: string;
+    author?: string;
+    msg?: string;
+}): request;
+export = _exports;