+++ workspace/packages/muse-core/types/storage/cache.d.ts
@@ -1,0 +1,2 @@
+export = cacheStorage;
+declare let cacheStorage: any;