+++ workspace/packages/muse-express-middleware/lib/app.js
@@ -1,0 +1,170 @@
+const _ = require('lodash');
+const museCore = require('@ebay/muse-core');
+const fs = require('fs');
+const logger = museCore.logger.createLogger('muse-express-middleware.app');
+const path = require('path');
+
+const defaultTemplate = `
+<!doctype html>
+<html lang="en">
+<head>
+  <title><%= title %></title>
+  <link rel="shortcut icon" href="<%= favicon %>" />
+  <script>
+    window.MUSE_GLOBAL = <%= museGlobal %>;
+  </script>
+</head>
+<body></body>
+<script src="<%= bootPluginUrl %>"></script>
+</html>
+`;
+
+const getAppInfoByUrl = async req => {
+  const appByUrl = await museCore.data.get('muse.app-by-url');
+  const host = req.get('host');
+  const fullPath = path.join(req.baseUrl || '/', req.path);
+  // fullurl example: 'www.example.com/foo/bar'
+  const fullUrl = host + fullPath;
+  let matchedUrl;
+  if (appByUrl[host]) {
+    matchedUrl = host;
+  } else {
+    matchedUrl = Object.keys(appByUrl).find(u => {
+      const uPath = u.endsWith('/') ? u : u + '/';
+      return (
+        // example.com equals example.com
+        u === host ||
+        // /foo equals /foo
+        u === fullPath ||
+        // example.com/foo/bar startsWith example.com/foo/
+        fullUrl.startsWith(uPath) ||
+        // /foo/bar startsWith /foo/
+        fullPath.startsWith(uPath)
+      );
+    });
+  }
+  if (!matchedUrl) {
+    return null;
+  }
+  return appByUrl[matchedUrl];
+};
+
+module.exports = ({
+  appName,
+  envName = 'staging',
+  cdn = '/muse-assets',
+  isDev = false,
+  isLocal = false,
+  template = defaultTemplate,
+  byUrl = false,
+  serviceWorker = '/sw.js', // If not use service worker, set it to false
+  serviceWorkerCacheName = 'muse_assets',
+}) => {
+  let swContent = fs.readFileSync(path.join(__dirname, './sw.js')).toString();
+  if (serviceWorkerCacheName) {
+    // Simply replace cacheName if necessary
+    // If your old cache has some issues, you may need to change cacheName to invalidate all old cache
+    swContent = swContent.replace(
+      "const cacheName = 'muse_assets';",
+      `const cacheName = '${serviceWorkerCacheName}';`,
+    );
+  }
+  return async (req, res, next) => {
+    if (serviceWorker && req.path === serviceWorker) {
+      // Service built-in service worker
+      res.set('Content-Type', 'application/javascript; charset=utf-8');
+      res.write(swContent);
+      res.end();
+      return;
+    }
+    const appInfo = museCore.plugin.invoke('museMiddleware.app.getAppInfo', req)[0];
+    if (appInfo) {
+      appName = appInfo.appName;
+      envName = appInfo.envName;
+    } else if (byUrl) {
+      const appInfo = await getAppInfoByUrl(req);
+      logger.info(`App info by url: ${JSON.stringify(appInfo)}`);
+
+      if (!appInfo) {
+        return next();
+      }
+      appName = appInfo.appName;
+      envName = appInfo.envName;
+    } else if (!appName) {
+      throw new Error(`No appName provided to start the Muse app.`);
+    } else {
+      logger.info(`Getting app by specified ${appName}/${envName}`);
+    }
+
+    logger.info(`Getting full app data: ${appName}`);
+    const app = await museCore.data.get(`muse.app.${appName}`);
+    if (!app) {
+      res.send('No app found: ' + appName);
+      return;
+    }
+    const env = app.envs?.[envName];
+    if (!env) {
+      res.send('No env found: ' + envName);
+      return;
+    }
+    const plugins = env.plugins;
+    museCore.plugin.invoke('museMiddleware.app.processPlugins', plugins, {
+      app,
+      appName,
+      envName,
+    });
+
+    const bootPlugins = plugins.filter(p => p.type === 'boot');
+
+    if (bootPlugins.length === 0) {
+      return res.send('No boot plugin.');
+    } else if (bootPlugins.length > 1) {
+      return res.send(
+        `There should be only one boot plugin, found ${bootPlugins.length}: ${bootPlugins.map(
+          p => p.name,
+        )}`,
+      );
+    }
+
+    const bootPlugin = bootPlugins[0];
+    const museGlobal = {
+      app: _.omit(app, ['envs']),
+      env: _.omit(env, ['plugins']),
+      appName: appName,
+      envName: envName,
+      plugins,
+      isDev,
+      isLocal,
+      cdn,
+      bootPlugin: bootPlugin.name,
+      // If app disabled service worker, or it's not confiugred for the app
+      serviceWorker:
+        !app.noServiceWorker && serviceWorker
+          ? path.join(req.baseUrl || '/', serviceWorker)
+          : false,
+    };
+
+    logger.info(`Muse global: ${JSON.stringify(museGlobal)}`);
+    museCore.plugin.invoke('museMiddleware.app.processMuseGlobal', museGlobal);
+    res.setHeader('Content-Type', 'text/html; charset=utf-8');
+
+    const favicon = app.iconId
+      ? `${cdn}/p/app-icon.${app.name}/v0.0.${app.iconId}/dist/icon.png`
+      : path.join(req.baseUrl || '/', 'favicon.png');
+    const ctx = {
+      indexHtml: _.template(template)({
+        title: app.title || 'Muse App',
+        favicon,
+        bootPluginUrl:
+          bootPlugin.url ||
+          `${cdn}/p/${museCore.utils.getPluginId(bootPlugin.name)}/v${
+            bootPlugin.version
+          }/dist/boot.js`,
+        museGlobal: JSON.stringify(museGlobal),
+      }),
+    };
+    museCore.plugin.invoke('museMiddleware.app.processIndexHtml', ctx);
+    res.write(ctx.indexHtml);
+    res.end();
+  };
+};